chore(security-rules): Refactor security rules error logic
