+++ src/security-rules/security-rules.ts
@@ -19,7 +19,7 @@ import * as validator from '../utils/validator';
 import {
   SecurityRulesApiClient, RulesetResponse, RulesetContent, ListRulesetsResponse,
 } from './security-rules-api-client-internal';
-import { FirebaseSecurityRulesError } from './security-rules-internal';
+import { FirebaseSecurityRulesError } from './error';
 
 /**
  * A source file containing some Firebase security rules. The content includes raw
