+++ test/unit/security-rules/security-rules-api-client.spec.ts
@@ -20,7 +20,7 @@ import * as _ from 'lodash';
 import * as chai from 'chai';
 import * as sinon from 'sinon';
 import { SecurityRulesApiClient, RulesetContent } from '../../../src/security-rules/security-rules-api-client-internal';
-import { FirebaseSecurityRulesError } from '../../../src/security-rules/security-rules-internal';
+import { FirebaseSecurityRulesError } from '../../../src/security-rules/error';
 import { HttpClient } from '../../../src/utils/api-request';
 import * as utils from '../utils';
 import * as mocks from '../../resources/mocks';
